# docs/openapi/mujian-relay.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "幕间模型网关 API",
    "version": "1.0.0",
    "description": "面向幕间用户的公开模型网关契约。所有需要鉴权的请求都使用 Authorization: Bearer <TOKEN>。在线调试会发起真实请求并可能消耗额度。"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前幕间网关"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/v1/models": {
      "get": {
        "tags": ["模型"],
        "summary": "列出可用模型",
        "operationId": "listOpenAIModels",
        "responses": {
          "200": { "$ref": "#/components/responses/ModelList" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/v1beta/models": {
      "get": {
        "tags": ["模型"],
        "summary": "列出 Gemini 模型",
        "operationId": "listGeminiModels",
        "responses": {
          "200": { "$ref": "#/components/responses/JSON" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1beta/openai/models": {
      "get": {
        "tags": ["模型"],
        "summary": "列出 Gemini OpenAI 兼容模型",
        "operationId": "listGeminiOpenAIModels",
        "responses": {
          "200": { "$ref": "#/components/responses/ModelList" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/v1/video/generations": {
      "post": {
        "tags": ["视频"],
        "summary": "创建兼容视频任务",
        "operationId": "createLegacyVideo",
        "requestBody": { "$ref": "#/components/requestBodies/LegacyVideo" },
        "responses": {
          "200": { "$ref": "#/components/responses/Task" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token",
        "description": "在文档顶部 Authorize 中手动粘贴 Token，不要包含 Bearer 前缀。"
      }
    },
...
    "requestBodies": {
...
      "LegacyVideo": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/LegacyVideoRequest" }
          }
        }
      },
...
    },
    "responses": {
      "JSON": {
        "description": "成功",
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
...
      "ModelList": {
        "description": "模型列表",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ModelList" }
          }
        }
      },
...
      "Task": {
        "description": "任务或任务结果",
        "content": {
          "application/json": { "schema": { "type": "object", "additionalProperties": true } }
        }
      },
...
      "Error": {
        "description": "请求失败",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        }
      }
    },
    "schemas": {
      "Model": {
        "type": "object",
        "required": ["id", "object"],
        "properties": {
          "id": { "type": "string" },
          "object": { "type": "string", "example": "model" },
          "created": { "type": "integer", "format": "int64" },
          "owned_by": { "type": "string" }
        }
      },
      "ModelList": {
        "type": "object",
        "required": ["object", "data"],
        "properties": {
          "object": { "type": "string", "example": "list" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Model" } }
        }
      },
...
      "LegacyVideoRequest": {
        "type": "object",
        "required": ["model", "prompt"],
        "properties": {
          "model": { "type": "string" },
          "prompt": { "type": "string" },
          "image": { "type": "string", "description": "图像 URL 或 Base64" },
          "images": { "type": "array", "items": { "type": "string" } },
          "size": { "type": "string" },
          "duration": { "type": "integer" },
          "seconds": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": true
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        },
        "additionalProperties": true
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "type": { "type": "string" },
          "param": { "type": "string", "nullable": true },
          "code": {
            "oneOf": [
              { "type": "string" },
              { "type": "integer" }
            ]
          }
        },
        "additionalProperties": true
      }
    }
  }
}